Type checking in Vector, Color, ProjectionSettings
Added statements that convert a number to float/int so that the Java implementation doesn't have to worry about them.

+++ python_api/src/api/geom/vector.py
@@ -4,10 +4,10 @@ from ..base import BaseAPIClass
 class Vector(BaseAPIClass):
 	"""Represents the berkerdemoglu.engine.graphics.math.Vector3D class."""
 
-	def __init__(self, x: int = 0, y: int = 0, z: int = 0):
-		self.x = x
-		self.y = y
-		self.z = z
+	def __init__(self, x: float = 0, y: float = 0, z: float = 0):
+		self.x = float(x)
+		self.y = float(y)
+		self.z = float(z)
 
 	def as_dict(self):
 		return self.__dict__
